:bug: fix detail map time selector and change date format
