+++ src/components/common/slider/DateSlider.tsx
@@ -1,0 +1,162 @@
+import React, { useCallback, useEffect, useState } from "react";
+import dayjs from "dayjs";
+import { Grid, Stack, Typography } from "@mui/material";
+import { dateToValue, valueToDate } from "../../../utils/DateUtils";
+import { dateDefault } from "../constants";
+import rc8Theme from "../../../styles/themeRC8";
+import { padding } from "../../../styles/constants";
+import PlainSlider from "./PlainSlider";
+
+interface DateSliderProps {
+  visible?: boolean;
+  currentMinDate: string | undefined;
+  currentMaxDate: string | undefined;
+  minDate: string;
+  maxDate: string;
+  onDateRangeChange: (
+    event: Event | React.SyntheticEvent<Element, Event>,
+    value: number | number[]
+  ) => void;
+}
+
+const COMPONENT_ID = "dateslider-daterange-menu-button";
+
+const DateSlider: React.FC<DateSliderProps> = ({
+  currentMinDate,
+  currentMaxDate,
+  minDate,
+  maxDate,
+  onDateRangeChange,
+}) => {
+  const [dateRangeStamp, setDateRangeStamp] = useState<number[]>([
+    dateToValue(
+      dayjs(currentMinDate ? currentMinDate : minDate, dateDefault.DATE_FORMAT)
+    ),
+    dateToValue(
+      dayjs(currentMaxDate ? currentMaxDate : maxDate, dateDefault.DATE_FORMAT)
+    ),
+  ]);
+
+  const handleSliderChange = useCallback(
+    (_: Event, newValue: number | number[]) => {
+      const v = newValue as number[];
+      setDateRangeStamp(v);
+    },
+    []
+  );
+
+  useEffect(() => {
+    const newDateRangeStamp = [
+      dateToValue(
+        dayjs(
+          currentMinDate ? currentMinDate : minDate,
+          dateDefault.DATE_FORMAT
+        )
+      ),
+      dateToValue(
+        dayjs(
+          currentMaxDate ? currentMaxDate : maxDate,
+          dateDefault.DATE_FORMAT
+        )
+      ),
+    ];
+    setDateRangeStamp(newDateRangeStamp);
+  }, [currentMinDate, currentMaxDate, minDate, maxDate]);
+
+  return (
+    <Grid
+      container
+      sx={{
+        backgroundColor: rc8Theme.palette.primary6,
+        borderRadius: "6px",
+        display: "flex",
+        width: "100%",
+        mx: "8px",
+      }}
+      data-testid={COMPONENT_ID}
+    >
+      <Grid
+        item
+        xs={12}
+        container
+        sx={{
+          px: padding.medium,
+          pt: { xs: "24px", md: padding.small },
+        }}
+      >
+        <Stack
+          width="100%"
+          direction="row"
+          alignItems="center"
+          mx={{ xs: "18px", sm: "6px" }}
+          gap="16px"
+        >
+          <Typography
+            sx={{
+              ...rc8Theme.typography.title1Medium,
+              color: rc8Theme.palette.text1,
+              whiteSpace: "nowrap",
+              mr: "8px",
+              display: { xs: "none", sm: "block" },
+            }}
+          >
+            Start Date
+          </Typography>
+          <PlainSlider
+            value={dateRangeStamp}
+            min={dateToValue(dayjs(minDate, dateDefault.DATE_FORMAT))}
+            max={dateToValue(dayjs(maxDate, dateDefault.DATE_FORMAT))}
+            onChangeCommitted={(_, value) => onDateRangeChange(_, value)}
+            onChange={handleSliderChange}
+            valueLabelDisplay="auto"
+            valueLabelFormat={(value: number) =>
+              valueToDate(value).format(dateDefault.DISPLAY_FORMAT)
+            }
+          />
+          <Typography
+            sx={{
+              ...rc8Theme.typography.title1Medium,
+              color: rc8Theme.palette.text1,
+              whiteSpace: "nowrap",
+              ml: "8px",
+              display: { xs: "none", sm: "block" },
+            }}
+          >
+            On going
+          </Typography>
+        </Stack>
+      </Grid>
+      <Grid
+        item
+        xs={12}
+        display="flex"
+        justifyContent="space-between"
+        alignItems="center"
+        sx={{
+          mx: "20px",
+          mt: { xs: "2px", md: "6px" },
+          mb: { xs: "2px", md: "6px" },
+        }}
+      >
+        <Typography
+          sx={{
+            ...rc8Theme.typography.body1Medium,
+            color: rc8Theme.palette.text1,
+          }}
+        >
+          {dayjs(minDate).format(dateDefault.DISPLAY_FORMAT)}
+        </Typography>
+        <Typography
+          sx={{
+            ...rc8Theme.typography.body1Medium,
+            color: rc8Theme.palette.text1,
+          }}
+        >
+          {dayjs(maxDate).format(dateDefault.DISPLAY_FORMAT)}
+        </Typography>
+      </Grid>
+    </Grid>
+  );
+};
+
+export default DateSlider;
